Extract Tag component
